+++ app/derivative_services/jp2_derivative_service.rb
@@ -45,13 +45,17 @@ class Jp2DerivativeService
   # @return [FileMetadata]
   def intermediate_target_files(resource)
     supported = resource.intermediate_files.select do |intermed|
-      ["image/tiff", "image/jpeg"].include?(intermed.mime_type.first)
+      valid_mime_types.include?(intermed.mime_type.first)
     end
     supported.empty? ? nil : supported.first
   end
 
   def valid?
-    ["image/tiff", "image/jpeg"].include?(mime_type.first)
+    valid_mime_types.include?(mime_type.first)
+  end
+
+  def valid_mime_types
+    ["image/tiff", "image/jpeg", "image/png"]
   end
 
   def create_derivatives
@@ -71,12 +75,7 @@ class Jp2DerivativeService
   end
 
   def run_derivatives
-    case mime_type
-    when ["image/tiff"]
-      run_tiff_derivatives
-    when ["image/jpeg"]
-      run_tiff_derivatives
-    end
+    run_tiff_derivatives if valid_mime_types.include?(mime_type.first)
   end
 
   def run_tiff_derivatives